+++ src/plugins/stake-plugins/generic/pluginInfo/cardanoKilnPool.ts
@@ -22,6 +22,7 @@ const kilnPolicyConfig: Array<StakePolicyConfig<CardanoPooledKilnAdapterConfig>>
       poolId: 'pool10rdglgh4pzvkf936p2m669qzarr9dusrhmmz9nultm3uvq4eh5k'
     },
     hideUnstakeAndClaimAction: true,
+    isLiquidStaking: true,
     stakeAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }],
     rewardAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }]
   },
@@ -44,6 +45,7 @@ const kilnPolicyConfig: Array<StakePolicyConfig<CardanoPooledKilnAdapterConfig>>
       poolId: 'pool1fcp4d2pxh0e7q5ju63sjqcdpxpr3pvxg6ykl23t6c97d7dnvjvw'
     },
     hideUnstakeAndClaimAction: true,
+    isLiquidStaking: true,
     stakeAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }],
     rewardAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }]
   },
@@ -66,6 +68,7 @@ const kilnPolicyConfig: Array<StakePolicyConfig<CardanoPooledKilnAdapterConfig>>
       poolId: 'pool1v62c7d92xv6gyh4x9rhfpkwzlpw2ypxk92xvzavakg3xypatklv'
     },
     hideUnstakeAndClaimAction: true,
+    isLiquidStaking: true,
     stakeAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }],
     rewardAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }]
   },
@@ -88,6 +91,7 @@ const kilnPolicyConfig: Array<StakePolicyConfig<CardanoPooledKilnAdapterConfig>>
       poolId: 'pool1mtxmk0skqkr5y0wxnxps4n35j6wn9q8dfr82y423vvlp53vccux'
     },
     hideUnstakeAndClaimAction: true,
+    isLiquidStaking: true,
     stakeAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }],
     rewardAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }]
   },
@@ -110,6 +114,7 @@ const kilnPolicyConfig: Array<StakePolicyConfig<CardanoPooledKilnAdapterConfig>>
       poolId: 'pool10d6mmw3mn9ku3r7uqqye672dz3sv76lh5kvh5rdpr9l5ug5yknr'
     },
     hideUnstakeAndClaimAction: true,
+    isLiquidStaking: true,
     stakeAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }],
     rewardAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }]
   },
@@ -132,6 +137,7 @@ const kilnPolicyConfig: Array<StakePolicyConfig<CardanoPooledKilnAdapterConfig>>
       poolId: 'pool1mtuhuh8hkf8am0qzx45y58kll8q83sjh6pwljrflcmw970d82f3'
     },
     hideUnstakeAndClaimAction: true,
+    isLiquidStaking: true,
     stakeAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }],
     rewardAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }]
   },
@@ -154,6 +160,7 @@ const kilnPolicyConfig: Array<StakePolicyConfig<CardanoPooledKilnAdapterConfig>>
       poolId: 'pool1aqg8vxzv75zhjzjjd9s20fu6r0xz70yl8lk3teacwy7qyc2p2j7'
     },
     hideUnstakeAndClaimAction: true,
+    isLiquidStaking: true,
     stakeAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }],
     rewardAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }]
   },
@@ -176,6 +183,7 @@ const kilnPolicyConfig: Array<StakePolicyConfig<CardanoPooledKilnAdapterConfig>>
       poolId: 'pool19kfm6lz5uw7nylq27swr367mqdycmug7tve94l6h3xsz64seqtc'
     },
     hideUnstakeAndClaimAction: true,
+    isLiquidStaking: true,
     stakeAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }],
     rewardAssets: [{ pluginId: 'cardano', currencyCode: 'ADA' }]
   }